+++ hub/src/server/drivers/diskDriver.js
@@ -4,7 +4,7 @@ import { BadPathError } from '../errors'
 import logger from 'winston'
 import Path from 'path'
 
-import type { DriverModel } from '../driverModel'
+import type { DriverModel, DriverStatics } from '../driverModel'
 import type { Readable } from 'stream'
 
 type DISK_CONFIG_TYPE = { diskSettings: { storageRootDirectory?: string },
@@ -15,6 +15,20 @@ class DiskDriver implements DriverModel {
   storageRootDirectory: string
   readURL: string
   pageSize: number
+
+
+  static getConfigInformation() {
+    const envVars = {}
+    if (process.env['GAIA_DISK_STORAGE_ROOT_DIR']) {
+      const diskSettings = { storageRootDirectory: process.env['GAIA_DISK_STORAGE_ROOT_DIR'] }
+      envVars['diskSettings'] = diskSettings
+    }
+
+    return {
+      defaults: { diskSettings: { storageRootDirectory: undefined } },
+      envVars
+    }
+  }
 
   constructor (config: DISK_CONFIG_TYPE) {
     if (!config.readURL) {
@@ -178,6 +192,8 @@ class DiskDriver implements DriverModel {
     })
   }
 }
+
+(DiskDriver: DriverStatics)
 
 module.exports = DiskDriver