+++ hub/CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for scoped authentication tokens via a new `scopes` field in
   the authentication JSON web token.
+- Configuration via environment variables. See the file `envvars.md` for
+  a listing.
+### Changed
+- The S3 and GCP drivers now respect the Cache-Control setting (the Azure
+  driver already did).
 
 ## [2.2.0]
 ### Added